Fix native part of ANT APIs to conform coding style

+++ api/antml/native/internal/ant_ml_native_internal.h
@@ -22,10 +22,10 @@ void ant_ml_getMaxOfBuffer_internal_uint8(const unsigned char *data_array,
                                           size_t data_array_len,
                                           int *result_max_index,
                                           unsigned char *result_value);
-void ant_ml_getMaxOfBuffer_internal_int32(const long *data_array,
+void ant_ml_getMaxOfBuffer_internal_int32(const int32_t *data_array,
                                           size_t data_array_len,
                                           int *result_max_index,
-                                          long *result_value);
+                                          int32_t *result_value);
 void ant_ml_getMaxOfBuffer_internal_float32(const float *data_array,
                                             size_t data_array_len,
                                             int *result_max_index,
