+++ api/antstream/native/internal/ant_stream_native_internal.h
@@ -20,10 +20,11 @@
 
 bool ant_stream_testPipeline_internal(const char *ipAddress);
 
-void ant_stream_callDbusMethod_internal(const char *inputMessage, char *resultMessage);
+void ant_stream_callDbusMethod_internal(const char *inputMessage,
+                                        char *resultMessage);
 void ant_stream_initializeStream_internal();
 
-typedef void (*ant_async_handler)(const char *, unsigned char *, unsigned long);
+typedef void (*ant_async_handler)(const char *, unsigned char *, uint32_t);
 bool ant_stream_elementConnectSignal_internal(int element_index,
                                               const char *detailed_signal,
                                               ant_async_handler handler);
